+++ views/js/front/applePayDirect/applePayDirectProduct.js
@@ -148,7 +148,7 @@ $(document).ready(function () {
                         ApplePaySession.STATUS_SUCCESS,
                         {
                             'amount': response.data.amount,
-                            'label': ' mollie'
+                            'label': totalLabel
                         },
                         []
                     )
@@ -191,7 +191,7 @@ $(document).ready(function () {
                             ApplePaySession.STATUS_SUCCESS,
                             response.shipping_methods,
                             {
-                                'label': firstTotal.label,
+                                'label': totalLabel,
                                 'amount': firstTotal.amount
                             },
                             [